refactor: 필터 직군 컴포넌트 리팩토링
