+++ src/components/home/portfolioList/HomeView/SearchBar.tsx
@@ -3,8 +3,8 @@ import { useState } from 'react';
 import { HomeProps } from 'HomeModule';
 import { Portfolio } from './Portfolio';
 import { useRecoilValue } from 'recoil';
-import { userInfoData, positionsData } from '../../../../utils/data/atom';
-import { Filter } from './Filter';
+import { userInfoData } from '../../../../utils/data/atom';
+import { FilterPosition } from './FilterPosition';
 
 export const SearchBar = ({
     filterBox,
@@ -26,27 +26,34 @@ export const SearchBar = ({
     const [searchValue, setSearchValue] = useState<string>('');
 
     const userInfo = useRecoilValue(userInfoData);
-    const positions = useRecoilValue(positionsData);
 
+    // 검색 로직
     const handleSearchSubmit = (e: React.SyntheticEvent) => {
         e.preventDefault();
+        // 필터 선택시 해당 하는 유저 정보
         const filteredUserInfo: Array<HomeProps.UserInfoProps> = userInfo.filter((item) => {
             if (selectedFilter.length === 0) return item;
             else if (selectedFilter.indexOf(item.job) >= 0) return item;
         });
+        // 검색 기능 이용 시 해당 하는 유저 정보
         const searchUserInfo: Array<HomeProps.UserInfoProps> = filteredUserInfo.filter((item) => {
             if (searchValue === item.name) return item;
             if (searchValue.toLowerCase() === item.user_skill.toLowerCase()) return item;
         });
+
+        // 필터를 선택한 경우
         if (selectedFilter.length > 0) {
+            // 검색 기능 이용 시 해당하는 유저가 없는 경우
             if (searchValue !== '' && searchUserInfo.length === 0) {
                 setUserPortfolio([]);
+                // 검색 기능 이용 시 헤딩하는 유저가 있는 경우
             } else if (searchValue !== '' && searchUserInfo.length > 0) {
                 setUserPortfolio(
                     searchUserInfo.map((item: HomeProps.UserInfoProps, idx: number) => {
                         return <Portfolio key={idx} {...item} />;
                     }),
                 );
+                // 검색 기능 이용하지 않은 경우
             } else if (searchValue === '') {
                 setUserPortfolio(
                     filteredUserInfo.map((item: HomeProps.UserInfoProps, idx: number) => {
@@ -54,15 +61,19 @@ export const SearchBar = ({
                     }),
                 );
             }
+            // 필터를 선택하지 않은 경우
         } else if (selectedFilter.length === 0) {
+            // 검색 기능 이용 시 유저가 있는 경우
             if (searchValue !== '' && searchUserInfo.length > 0) {
                 setUserPortfolio(
                     searchUserInfo.map((item: HomeProps.UserInfoProps, idx: number) => {
                         return <Portfolio key={idx} {...item} />;
                     }),
                 );
+                // 검색 기능 이용 시 유저가 없는 경우
             } else if (searchValue !== '' && searchUserInfo.length === 0) {
                 setUserPortfolio([]);
+                // 검색 기능도 이용하지 않은 경우
             } else if (searchValue === '') {
                 setUserPortfolio(
                     filteredUserInfo.map((item: HomeProps.UserInfoProps, idx: number) => {
@@ -75,25 +86,13 @@ export const SearchBar = ({
 
     return (
         <SearchDiv>
-            {/* Position */}
-            <FilterPositionDiv>
-                <FilterButton
-                    onClick={() => {
-                        positionActive ? setPositionActive(false) : setPositionActive(true);
-                    }}
-                    isActive={positionActive}
-                >
-                    직군
-                </FilterButton>
-                {positionActive && (
-                    <Filter
-                        data={positions}
-                        filterBox={filterBox}
-                        selectedFilter={selectedFilter}
-                        setSelectedFilter={setSelectedFilter}
-                    />
-                )}
-            </FilterPositionDiv>
+            <FilterPosition
+                filterBox={filterBox}
+                positionActive={positionActive}
+                setPositionActive={setPositionActive}
+                selectedFilter={selectedFilter}
+                setSelectedFilter={setSelectedFilter}
+            />
             <SearchForm onSubmit={handleSearchSubmit}>
                 <SearchInput
                     type="text"
@@ -129,12 +128,6 @@ const SearchDiv = styled.div`
     display: flex;
     height: 36px;
     margin-bottom: 12px;
-`;
-const FilterPositionDiv = styled.div`
-    width: 30%;
-    position: relative;
-    box-sizing: border-box;
-    padding-right: 20px;
 `;
 const SearchForm = styled.form`
     box-sizing: border-box;
@@ -176,21 +169,5 @@ const ResetButton = styled.button`
 
     &:hover {
         opacity: 0.7;
-    }
-`;
-const FilterButton = styled.button<HomeProps.IFilterProps>`
-    width: 100%;
-    height: 44px;
-    border: 1px solid #e0e0e0;
-    padding: 10px;
-    color: ${(props) => (props.isActive ? props.theme.color.defaultText : props.theme.color.buttonText)};
-    border-radius: 4px;
-    font-size: 16px;
-    text-align: left;
-    background-color: ${(props) => (props.isActive ? props.theme.color.buttonBackground : props.theme.color.sub)};
-
-    &:hover {
-        color: ${(props) => props.theme.color.defaultText};
-        background-color: ${(props) => props.theme.color.buttonBackground};
     }
 `;
